Split #[derive(DeserializeEnum)] macro
