+++ macros/ast_node/src/lib.rs
@@ -14,6 +14,7 @@ use swc_macros_common::prelude::*;
 use syn::*;
 
 mod ast_node_macro;
+mod enum_deserialize;
 mod fold;
 mod from_variant;
 mod spanned;
@@ -87,7 +88,7 @@ pub fn derive_deserialize_enum(input: proc_macro::TokenStream) -> proc_macro::To
     let input = parse::<DeriveInput>(input).expect("failed to parse input as DeriveInput");
 
     let item =
-        ast_node_macro::expand_enum(input)
+        enum_deserialize::expand(input)
             .into_iter()
             .fold(TokenStream::new(), |mut t, item| {
                 item.to_tokens(&mut t);
